requirements.txt
streamlit>=1.0
tensorflow>=2.3
numpy>=1.19
pillow>=8.0